% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parseSPD}
\alias{parseSPD}
\title{SPD parser}
\usage{
parseSPD(filename, encoding = "latin1")
}
\arguments{
\item{filename}{path to file (SPD-format)}

\item{encoding}{encoding for the file}
}
\value{
list with elements:
\describe{
\item{Sform}{S-form (stations / hauls)}
\item{Tform}{T-form (catch samples)}
\item{Uform}{U-form (length frequencies (fish))}
\item{Vform}{V-form (individual measurements (fish))}
\item{Wform}{W-form (length frequencies / devstage frequencies (prey))}
}
}
\description{
Parses SPD froms (spesified in SPD handbook jan. 2010)
}
\details{
Units are only standardized when interpretation does not depend on other fields.
Incomplete forms will produce warnings (parsing failures). Missing columns will be interpreted as NAs
}
