% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readSform}
\alias{readSform}
\title{Parses all S-lines from file (spesified in SPD handbook jan. 2010)}
\usage{
readSform(file)
}
\arguments{
\item{file}{path to file (SPD-format)}
}
\value{
data.table
}
\description{
Parses all S-lines from file (spesified in SPD handbook jan. 2010)
}
