% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readTform}
\alias{readTform}
\title{Parses all T-lines from file (spesified in SPD handbook (missing mage, genetikk))}
\usage{
readTform(file)
}
\arguments{
\item{file}{path to file (SPD-format)}
}
\value{
data.table
}
\description{
Parses all T-lines from file (spesified in SPD handbook (missing mage, genetikk))
}
