% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readUform}
\alias{readUform}
\title{Parses all U-lines from file (spesified in SPD handbook jan. 2010)}
\usage{
readUform(file)
}
\arguments{
\item{file}{path to file (SPD-format)}
}
\value{
data.table
}
\description{
Parses all U-lines from file (spesified in SPD handbook jan. 2010)
}
