% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{readWform}
\alias{readWform}
\title{Parses all V-lines from file (spesified in SPD handbook jan. 2010)}
\usage{
readWform(file)
}
\arguments{
\item{file}{path to file (SPD-format)}
}
\value{
data.table
}
\description{
Parses all V-lines from file (spesified in SPD handbook jan. 2010)
}
